thinkpad_acpi: battery: Add inhibit_charge_minutes attribute
This should be considered experimental. There seem to be to ways to
inhibit charging, one globally using peak shift state, and one using
a battery-specific inhibit charge command. The latter is implemented
here. Setting a permanent timeout is not supported currently.

Reading the timeout is not supported because the ACPI call needed
to read the value is not documented/reverse-engineered yet.

v2: Add support for reading the attribute

Signed-off-by: Julian Andres Klode <jak@jak-linux.org>